Moved api service to new _services folder.
