+++ src/app/components/farmer/equipment-create/equipment-create.component.ts
@@ -6,7 +6,7 @@ import { PortfolioService } from '../../../services/portfolio/portfolio.service'
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
-import { ApiService } from 'src/app/services/api/api.service';
+import { ApiService } from 'src/app/_services/api-service/api.service';
 
 @Component({
     selector: 'app-equipment-create',
