+++ src/app/components/farmer/equipment-edit/equipment-edit.component.ts
@@ -7,7 +7,7 @@ import {
 } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Asset } from 'src/app/_models/asset';
-import { ApiService } from 'src/app/services/api/api.service';
+import { ApiService } from 'src/app/_services/api-service/api.service';
 import { PortfolioService } from 'src/app/services/portfolio/portfolio.service';
 
 @Component({
