+++ src/app/pages/bookkeeping/bookkeeping-view-record/bookkeeping-view-record.component.ts
@@ -21,7 +21,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { BookkeepingDeleteModalContentComponent } from 'src/app/components/modal/bookkeeping-delete-modal-content/bookkeeping-delete-modal-content.component';
 import { IncomeStatementItem } from 'src/app/_models/IncomeStatementItem';
-import { ApiService } from 'src/app/services/api/api.service';
+import { ApiService } from 'src/app/_services/api-service/api.service';
 import { BookkeepingService } from 'src/app/services/bookkeeping/bookkeeping.service';
 
 @Component({
