+++ src/app/services/authorization-services/authentication/auth.service.ts
@@ -18,7 +18,7 @@
 
 import { Injectable } from '@angular/core';
 import { JWTTokenService } from '../JWT-token-service/jwt-token.service';
-import { ApiService } from '../../api/api.service';
+import { ApiService } from '../../../_services/api-service/api.service';
 import { Router } from '@angular/router';
 import { BehaviorSubject, Observable, Subscription } from 'rxjs';
